+++ src/interfaces/event-flow/AccountCondition.ts
