+++ src/interfaces/index.ts
@@ -14,4 +14,4 @@ export * from './NetworkMap';
 export * from './RedisConfig';
 export * from './event-flow/ConditionEdge';
 export * from './event-flow/EntityCondition';
-export * from './event-flow/AccountCondtion';
+export * from './event-flow/AccountCondition';
